+++ CHANGELOG.md
@@ -155,6 +155,8 @@ All notable changes to this project will be documented in this file.
 
 - Refactored validate_providers_api_raw to use PointBlank, and added the new validation script `validate_cqc_providers_1_delta_api_download.py` with its corresponding validation test.
 
+- Changed references of domain=CQC_delta to domain=CQC. The bulk download pipeline is no longer used, the delta pipeline has taken it's place.
+
 - Changed the remove_duplicate_cqc_care_homes function as followed:
   - changed function name to remove_dual_registration_cqc_care_homes.
   - updated doc string with information from CQC.